Refactored the interfaces
